+++ compiler/compiler_utils/src/utils/mod.rs
@@ -1,2 +1,3 @@
 pub mod map;
 pub mod num;
+pub mod lifecycles;
