--- compiler/pastel_ir/Cargo.toml
@@ -1,6 +1,0 @@
-[package]
-name = "pastel_ir"
-version = "0.1.0"
-edition = "2024"
-
-[dependencies]
