--- compiler/pastel_parser/Cargo.toml
@@ -1,6 +1,0 @@
-[package]
-name = "pastel_parser"
-version = "0.1.0"
-edition = "2024"
-
-[dependencies]
