--- ir/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "ir"
-version = "0.1.0"
-edition = "2024"
-
-[dependencies]
-inkwell = { version = "0.8.0", features = ["llvm21-1"]}
-commons = { path = "../commons" }
-lexer = { path = "../lexer" }
-parser = { path = "../parser" }
-errors = { path = "../errors" }
