--- ir/src/irstruct/mod.rs
@@ -1,6 +1,0 @@
-//! Common definitions used internally to structure the IR.
-
-pub mod ptr;
-pub mod funcs;
-pub mod staticvars;
-pub mod structs;
