--- ir/src/lib.rs
@@ -1,12 +1,0 @@
-//! Quickfall's Intermediate Representation code. Converts the AST into instruction like nodes to pass down to compilation.
-
-
-pub mod types;
-pub mod values;
-pub mod math;
-pub mod irstruct;
-pub mod conv;
-pub mod ctx;
-pub mod refs;
-pub mod bools;
-pub mod utils;
