--- ir/src/types/mod.rs
@@ -1,22 +1,0 @@
-//! Quickfall's IR typing system. Basically contains any kind of code related to the types in Quickfall to an IR level.
-
-pub const SIGNED8_TYPE_HASH: u64 = 14203960237928745581;
-pub const SIGNED16_TYPE_HASH: u64 = 15304920686014172958;
-pub const SIGNED32_TYPE_HASH: u64 = 11583643502875509928;
-pub const SIGNED64_TYPE_HASH: u64 = 7572830400006405400;
-pub const SIGNED128_TYPE_HASH: u64 = 4344713510544482692;
-
-pub const UNSIGNED8_TYPE_HASH: u64 = 5290224216008800543;
-pub const UNSIGNED16_TYPE_HASH: u64 = 3436424547679103825;
-pub const UNSIGNED32_TYPE_HASH: u64 = 2263434253207377038;
-pub const UNSIGNED64_TYPE_HASH: u64 = 9321926220436863523;
-pub const UNSIGNED128_TYPE_HASH: u64 = 9667664784078163512;
-
-pub const POINTER_TYPE_HASH: u64 = 9259207781742941243;
-
-pub const BOOL_TYPE_HASH: u64 = 10699440025065254663;
-
-pub const STATICSTR_TYPE_HASH: u64 = 10484269008925909607;
-
-pub mod typing;
-pub mod storage;
