--- parser/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "parser"
-version = "0.1.0"
-edition = "2024"
-
-[dependencies]
-utils = { path = "../utils" }
-lexer = { path = "../lexer" }
-commons = { path = "../commons" }
-errors = { path = "../errors" }
