--- parser/src/ast/control/forloop.rs
@@ -1,35 +1,0 @@
-use errors::errs::CompilerResult;
-use lexer::token::{LexerToken, LexerTokenType};
-
-use crate::ast::{func::parse_node_body, parse_ast_node_in_body, parse_ast_value, tree::{ASTTreeNode, ASTTreeNodeKind}, var::decl::parse_variable_declaration};
-
-pub fn parse_for_loop(tokens: &Vec<LexerToken>, ind: &mut usize) -> CompilerResult<Box<ASTTreeNode>> {
-	let start = tokens[*ind].pos.clone();
-
-	*ind += 1;
-
-	tokens[*ind].expects(LexerTokenType::ParenOpen)?;
-
-	let initial = parse_variable_declaration(tokens, ind)?;
-
-	tokens[*ind].expects(LexerTokenType::Comma)?;
-
-	*ind += 1;
-	let cond = parse_ast_value(tokens, ind)?;
-
-	tokens[*ind].expects(LexerTokenType::Comma)?;
-	*ind += 1;
-
-	let increment = parse_ast_node_in_body(tokens, ind)?;
-
-	tokens[*ind].expects(LexerTokenType::ParenClose)?;
-	*ind += 1;
-
-	tokens[*ind].expects(LexerTokenType::BracketOpen)?;
-
-	let body = parse_node_body(tokens, ind)?;
-
-	let end = tokens[*ind - 1].get_end_pos();
-
-	return Ok(Box::new(ASTTreeNode::new(ASTTreeNodeKind::ForBlock { initial_state: initial, cond, increment, body }, start, end)));
-}
