--- parser/src/ast/control/mod.rs
@@ -1,8 +1,0 @@
-//!
-//! AST Parsing for any control kind of elements (eg loops, if statements, ...)
-//! 
-//! 
-
-pub mod ifelse;
-pub mod whileblock;
-pub mod forloop;
