--- parser/src/ast/control/whileblock.rs
@@ -1,23 +1,0 @@
-use errors::errs::CompilerResult;
-use lexer::token::LexerToken;
-
-use crate::ast::{control::ifelse::parse_condition_member, func::parse_node_body, tree::{ASTTreeNode, ASTTreeNodeKind}};
-
-pub fn parse_while_block(tokens: &Vec<LexerToken>, ind: &mut usize) -> CompilerResult<Box<ASTTreeNode>> {
-	let start = tokens[*ind].pos.clone();
-
-	*ind += 1;
-
-	let cond = parse_condition_member(tokens, ind)?;
-
-	tokens[*ind].expects(lexer::token::LexerTokenType::BracketOpen)?;
-
-	let body = match parse_node_body(tokens, ind) {
-		Ok(v) => v,
-		Err(e) => return Err(e)
-	};
-
-	let end = tokens[*ind - 1].get_end_pos().clone();
-
-	return Ok(Box::new(ASTTreeNode::new(ASTTreeNodeKind::WhileBlock { cond, body }, start, end)));
-}
