--- parser/src/ast/var/mod.rs
@@ -1,4 +1,0 @@
-//! AST parsing for variable related elements (variable declarations, modifications, ...)
-
-pub mod decl;
-pub mod staticdecl;
